refactor: reorganize password management decorators into one decorator object
